fix(pruning): address PR review feedback
- Move get_default_answer_from_graph() to context.py for modularity
  (Gemini review: function is a generic graph query, not pruning-specific)
- Guard against missing/empty raw_id in graph lookup — return None
  instead of empty string (Claude review: silent bug risk)
- Add clarifying comment for None canonical case in _get_noncanonical_answers
- Remove local import in dilemma_scoring.py — now imports from context.py
  at module level (no circular dependency)
